+++ src/lib.rs
@@ -12,6 +12,7 @@
     )
 )]
 
+mod analyze;
 pub mod app;
 mod cli;
 mod journal;